Todoアプリ作成
チェックボックス、ラジオボタン修正
